Initial draft of abl-imatinib-explicit example
