+++ examples/README.md
@@ -1,0 +1,4 @@
+# Examples
+
+* `abl-imatinib-explicit/` - alchemical relaxation of imatinib in Abl binding site
+* `loop-refinement/` - alchemical loop refinement
