+++ examples/abl-imatinib-explicit/README.md
@@ -1,0 +1,26 @@
+# Alchemical softening of imatinib bound to Abl
+
+This example uses fused alchemical softcore potentials to relax the geometry of imatinib in the binding site of Abl kinase.
+
+## Manifest
+
+* `soften-ligand.py` - example driver script for alchemically softening the ligand
+* `setup/` - these files were automatically generated by YANK; they contain Amber `prmtop/inpcrd` files for the solvated complex, used here
+
+## Usage
+
+Make sure you have the omnia channel and `dev` label added to your conda path:
+```bash
+# Add omnia channel for stable omnia packages
+- conda config --add channels omnia
+# Add omnia dev channels to pick up `openmm 7.1.0` pre-release
+- conda config --add channels omnia/label/dev
+```
+Install SAMS:
+```bash
+conda install --yes sams
+```
+Run the example:
+```bash
+python soften-ligand.py
+```
