+++ examples/abl-imatinib-explicit/setup/molecules/STI/STI.gaff.mol2
@@ -1,0 +1,153 @@
+@<TRIPOS>MOLECULE
+MOL
+   69    73     1     0     0
+SMALL
+No Charge or Current Charge
+
+
+@<TRIPOS>ATOM
+      1 C1         22.8270   11.7250   13.7800 ca        1 MOL     -0.255632
+      2 C2         22.0360   12.1770   14.8390 ca        1 MOL     -0.047601
+      3 C3         20.6680   11.8090   14.9590 cp        1 MOL     -0.283801
+      4 C4         20.1400   11.0010   13.9470 ca        1 MOL      0.420599
+      5 N1         20.8920   10.5700   12.9040 nb        1 MOL     -0.665801
+      6 C5         22.1760   10.9310   12.8420 ca        1 MOL      0.413299
+      7 C6         19.7780   12.2610   16.1190 cp        1 MOL      0.548299
+      8 C7         20.2660   12.8280   17.2280 ca        1 MOL     -0.405401
+      9 C8         19.3470   13.1770   18.2350 ca        1 MOL      0.460699
+     10 N2         18.0470   12.9440   18.0700 nb        1 MOL     -0.739701
+     11 C9         17.5930   12.3830   16.9620 ca        1 MOL      0.885999
+     12 N3         18.4330   12.0210   15.9770 nb        1 MOL     -0.761601
+     13 N4         16.2540   12.1250   16.6950 nh        1 MOL     -0.745401
+     14 C10        15.3460   12.2710   17.7100 ca        1 MOL      0.175899
+     15 C11        15.2340   11.3050   18.7320 ca        1 MOL     -0.106501
+     16 C12        14.2850   11.5390   19.7140 ca        1 MOL     -0.098101
+     17 C13        13.4510   12.6990   19.6990 ca        1 MOL     -0.148301
+     18 C14        13.5600   13.6330   18.6980 ca        1 MOL      0.051599
+     19 C15        14.5060   13.4010   17.6880 ca        1 MOL     -0.159001
+     20 N5         12.6590   14.7350   18.8140 n         1 MOL     -0.453001
+     21 C16        12.5890   15.6820   17.8510 c         1 MOL      0.669199
+     22 C17        11.4550   16.6420   17.9000 ca        1 MOL     -0.103901
+     23 C18        11.0790   17.3300   16.7130 ca        1 MOL     -0.078401
+     24 C19        10.0060   18.2850   16.6780 ca        1 MOL     -0.122701
+     25 C20         9.2980   18.5360   17.8690 ca        1 MOL     -0.153401
+     26 C21         9.6620   17.8540   19.0650 ca        1 MOL     -0.122701
+     27 C22        10.7360   16.9080   19.1120 ca        1 MOL     -0.078401
+     28 C23         8.1960   19.5290   17.8320 c3        1 MOL      0.169699
+     29 N6          8.1910   20.4060   18.9880 n4        1 MOL     -0.688101
+     30 C24         9.4550   21.1520   19.1060 c3        1 MOL      0.108999
+     31 C25         9.4480   22.0490   20.4020 c3        1 MOL      0.142899
+     32 N7          8.1500   22.8120   20.4470 n3        1 MOL     -0.741401
+     33 C26         8.1830   23.6190   21.6400 c3        1 MOL      0.152699
+     34 C27         6.9280   22.0290   20.3800 c3        1 MOL      0.142899
+     35 C28         6.9520   21.2300   19.0530 c3        1 MOL      0.108999
+     36 O1         13.3990   15.8030   16.9310 o         1 MOL     -0.571701
+     37 C29        16.0750   10.0530   18.8150 c3        1 MOL     -0.055001
+     38 H1         23.8740   11.9750   13.6950 ha        1 MOL      0.149799
+     39 H2         22.4750   12.8220   15.5850 ha        1 MOL      0.156099
+     40 H3         19.1010   10.7110   13.9980 h4        1 MOL      0.024199
+     41 H4         22.7530   10.5800   11.9990 h4        1 MOL      0.029799
+     42 H5         21.3240   13.0110   17.3460 ha        1 MOL      0.160999
+     43 H6         19.6990   13.6360   19.1470 h4        1 MOL      0.028299
+     44 H7         15.9450   11.8120   15.6870 hn        1 MOL      0.457299
+     45 H8         14.1720   10.8230   20.5150 ha        1 MOL      0.144299
+     46 H9         12.7250   12.8430   20.4850 ha        1 MOL      0.172799
+     47 H10        14.5920   14.1060   16.8740 ha        1 MOL      0.145499
+     48 H11        12.0080   14.8230   19.6970 hn        1 MOL      0.320399
+     49 H12        11.6220   17.1260   15.8020 ha        1 MOL      0.172699
+     50 H13         9.7490   18.7970   15.7620 ha        1 MOL      0.139599
+     51 H14         9.1080   18.0580   19.9690 ha        1 MOL      0.139599
+     52 H15        10.9970   16.4090   20.0340 ha        1 MOL      0.172699
+     53 H16         8.2980   20.1570   16.9340 hx        1 MOL      0.107499
+     54 H17         7.2300   19.0040   17.8040 hx        1 MOL      0.107499
+     55 H18         8.1630   19.7680   19.8840 hn        1 MOL      0.444299
+     56 H19        10.2940   20.4440   19.1660 hx        1 MOL      0.109599
+     57 H20         9.5830   21.7980   18.2250 hx        1 MOL      0.109599
+     58 H21         9.5380   21.4090   21.2920 h1        1 MOL      0.074799
+     59 H22        10.2950   22.7500   20.3670 h1        1 MOL      0.074799
+     60 H23         7.2510   24.1980   21.7170 h1        1 MOL      0.057499
+     61 H24         9.0400   24.3080   21.5930 h1        1 MOL      0.057499
+     62 H25         8.2840   22.9680   22.5210 h1        1 MOL      0.057499
+     63 H26         6.8850   21.3430   21.2390 h1        1 MOL      0.074799
+     64 H27         6.0590   22.7030   20.4050 h1        1 MOL      0.074799
+     65 H28         6.9320   21.9300   18.2050 hx        1 MOL      0.109599
+     66 H29         6.0710   20.5720   19.0060 hx        1 MOL      0.109599
+     67 H30        15.7860    9.4760   19.7060 hc        1 MOL      0.050099
+     68 H31        17.1370   10.3290   18.8860 hc        1 MOL      0.050099
+     69 H32        15.9140    9.4430   17.9140 hc        1 MOL      0.050099
+@<TRIPOS>BOND
+     1    1    2 ar  
+     2    1    6 ar  
+     3    1   38 1   
+     4    2    3 ar  
+     5    2   39 1   
+     6    3    4 ar  
+     7    3    7 1   
+     8    4    5 ar  
+     9    4   40 1   
+    10    5    6 ar  
+    11    6   41 1   
+    12    7    8 ar  
+    13    7   12 ar  
+    14    8    9 ar  
+    15    8   42 1   
+    16    9   10 ar  
+    17    9   43 1   
+    18   10   11 ar  
+    19   11   12 ar  
+    20   11   13 1   
+    21   13   14 1   
+    22   13   44 1   
+    23   14   15 ar  
+    24   14   19 ar  
+    25   15   16 ar  
+    26   15   37 1   
+    27   16   17 ar  
+    28   16   45 1   
+    29   17   18 ar  
+    30   17   46 1   
+    31   18   19 ar  
+    32   18   20 1   
+    33   19   47 1   
+    34   20   21 1   
+    35   20   48 1   
+    36   21   22 1   
+    37   21   36 2   
+    38   22   23 ar  
+    39   22   27 ar  
+    40   23   24 ar  
+    41   23   49 1   
+    42   24   25 ar  
+    43   24   50 1   
+    44   25   26 ar  
+    45   25   28 1   
+    46   26   27 ar  
+    47   26   51 1   
+    48   27   52 1   
+    49   28   29 1   
+    50   28   53 1   
+    51   28   54 1   
+    52   29   30 1   
+    53   29   35 1   
+    54   29   55 1   
+    55   30   31 1   
+    56   30   56 1   
+    57   30   57 1   
+    58   31   32 1   
+    59   31   58 1   
+    60   31   59 1   
+    61   32   33 1   
+    62   32   34 1   
+    63   33   60 1   
+    64   33   61 1   
+    65   33   62 1   
+    66   34   35 1   
+    67   34   63 1   
+    68   34   64 1   
+    69   35   65 1   
+    70   35   66 1   
+    71   37   67 1   
+    72   37   68 1   
+    73   37   69 1   
+@<TRIPOS>SUBSTRUCTURE
+     1 MOL         1 TEMP              0 ****  ****    0 ROOT
